+++ frontend/src/component/impact-views/views/GoalSummaryPanel/sparkline.test.ts
