+++ frontend/src/component/impact-views/views/GoalSummaryPanel/sparkline.ts
